requirements.txt
alembic==1.7.4
autopep8==1.6.0
Babel==2.9.0
click==8.0.3
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-Moment==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2==2.9.1
pycodestyle==2.8.0
python-dateutil==2.6.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
Werkzeug==2.0.2
WTForms==2.3.3
